# swagger/crowd-lending.yml
swagger: '2.0'
info:
  version: '1.0'
  title: Crowd Lending
  description: API Suite for Crowd Lending platforms
host: 'localhost:8080'
basePath: /v1
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Authorization:
    name: Authorization
    type: apiKey
    in: header
  X-API-KEY:
    name: X-API-KEY
    type: apiKey
    in: header
paths:
...
  '/borrowers/{borrower_id}':
    parameters:
      - name: borrower_id
        in: path
        required: true
        type: string
    get:
      operationId: get-borrower
      summary: Get Borrower
      tags:
        - Borrowers
      responses:
        '200':
          description: 'Successful Response'
          schema:
            $ref: '#/definitions/borrower-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
    put:
      operationId: update-borrower
      summary: Update Borrower
      tags:
        - Borrowers
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/borrower'
      responses:
        '200':
          description: 'Successful Response'
          schema:
            $ref: '#/definitions/borrower-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
    delete:
      operationId: delete-borrower
      summary: Delete Borrower
      tags:
        - Borrowers
      responses:
        '204':
          description: 'Successfully deleted'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
  /borrowers:
    get:
      operationId: list-borrowers
      summary: List Borrowers
      tags:
        - Borrowers
      parameters:
        - name: name
          in: query
          description: Search all borrowers by name
          type: string
        - name: email
          in: query
          description: Search all borrowers by email
          type: string
      responses:
        '200':
          description: 'Successful Response'
          schema:
            type: array
            items:
              $ref: '#/definitions/borrower-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
    post:
      operationId: create-borrower
      summary: Create Borrower
      tags:
        - Borrowers
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/borrower'
      responses:
        '201':
          description: 'Successfully created Borrower'
          schema:
            $ref: '#/definitions/borrower-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
  '/investors/{investor_id}':
    parameters:
      - name: investor_id
        in: path
        required: true
        type: string
    get:
      operationId: get-investor
      summary: Get Investor
      tags:
        - Investors
      responses:
        '200':
          description: 'Successful Response'
          schema:
            $ref: '#/definitions/investor-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
    put:
      operationId: Update-investor
      summary: Update Investor
      tags:
        - Investors
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/investor'
      responses:
        '200':
          description: 'Successful Response'
          schema:
            $ref: '#/definitions/investor-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
    delete:
      operationId: delete-investor
      summary: Delete Investor
      tags:
        - Investors
      responses:
        '204':
          description: 'Successfully Deleted'
      security:
        - Authorization: []
        - X-API-KEY: []
  /investors:
    get:
      operationId: list-investors
      summary: List Investors
      tags:
        - Investors
      parameters:
        - name: name
          in: query
          description: Search all investors by name
          type: string
        - name: email
          in: query
          description: Search all investors by email
          type: string
      responses:
        '200':
          description: 'Successful Response'
          schema:
            type: array
            items:
              $ref: '#/definitions/investor-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
    post:
      operationId: create-investor
      summary: Create Investor
      tags:
        - Investors
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/investor'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/investor-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
definitions:
  investor:
    title: Investor
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      phone_number:
        type: string
      national_identifier:
        type: string
      address:
        type: string
      registration_date:
        type: string
  borrower:
    title: Borrower
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      phone_number:
        type: string
      national_identifier:
        type: string
      address:
        type: string
      registration_date:
        type: string
        format: date
      borrower_account:
        $ref: '#/definitions/account-details'
...
  server-error:
    title: Server Error
    type: object
    properties:
      code:
        type: string
      message:
        type: string
      tracking_id:
        type: string
  generic-error:
    title: Generic Error
    type: object
    properties:
      code:
        type: string
      message:
        type: string
...
  borrower-info:
    title: Borrower Info with ID
    type: object
    properties:
      id:
        type: string
      borrower:
        $ref: '#/definitions/borrower'
  investor-info:
    title: Investor Info with ID
    type: object
    properties:
      id:
        type: string
      investor:
        $ref: '#/definitions/investor'
      investments:
        type: array
        items:
          $ref: '#/definitions/investment-loan'
  investment-loan:
    title: Investment of a loan
    type: object
    properties:
      loan_id:
        type: string
      investment_id:
        type: string
      amount:
        type: number
      currency:
        type: string
      payment_transaction_id:
        type: string
      payment_date:
        type: string
        format: date
...
  account-details:
    title: Account Details
    type: object
    properties:
      bank_name:
        type: string
      bank_swift_code:
        type: string
      account_number:
        type: string
      account_name:
        type: string